+++ ext_emconf.php
@@ -12,10 +12,10 @@ $EM_CONF[$_EXTKEY] = [
     'uploadFolder' => false,
     'createDirs' => '',
     'clearCacheOnLoad' => true,
-    'version' => '1.3.0',
+    'version' => '2.0.0',
     'constraints' => [
         'depends' => [
-            'typo3' => '8.7.0-9.5.999',
+            'typo3' => '9.5.999',
         ],
         'conflicts' => [],
         'suggests' => [],