+++ ext_localconf.php
@@ -3,26 +3,23 @@
 defined('TYPO3_MODE') || die('Access denied.');
 
 call_user_func(function () {
-    $defaultConfiguration = [
-        'hashBaseCustomization' => [
-            'enabled' => true,
-        ],
-    ];
-
     $GLOBALS['TYPO3_CONF_VARS']['EXT']['EXTCONF']['singleview']
         = array_replace_recursive(
         !empty($GLOBALS['TYPO3_CONF_VARS']['EXT']['EXTCONF']['singleview'])
             ? $GLOBALS['TYPO3_CONF_VARS']['EXT']['EXTCONF']['singleview'] : [],
-        $defaultConfiguration
+        // default config
+        [
+            'hashBaseCustomization' => [
+                'enabled' => true,
+            ],
+        ]
     );
 
-    $singleViewConf = $GLOBALS['TYPO3_CONF_VARS']['EXT']['EXTCONF']['singleview'];
+    $GLOBALS['TYPO3_CONF_VARS']['SC_OPTIONS']['tslib/class.tslib_fe.php']['createHashBase']['singleview'] =
+        \SourceBroker\Singleview\Hooks\SingleViewPagePathLogic::class . '->init';
 
-    if ($singleViewConf['hashBaseCustomization']['enabled']) {
+    if ($GLOBALS['TYPO3_CONF_VARS']['EXT']['EXTCONF']['singleview']['hashBaseCustomization']['enabled']) {
         $GLOBALS['TYPO3_CONF_VARS']['SC_OPTIONS']['tslib/class.tslib_fe.php']['createHashBase']['singleview_hashBaseCustomization'] =
             \SourceBroker\Singleview\Hooks\HashBase::class . '->init';
     }
-
-    $GLOBALS['TYPO3_CONF_VARS']['SC_OPTIONS']['tslib/class.tslib_fe.php']['configArrayPostProc'][] =
-        \SourceBroker\Singleview\Hooks\SingleViewPagePathLogic::class . '->init';
 });