+++ CHANGELOG.rst
@@ -1,8 +1,8 @@
 Changelog
 ---------
 
-master
-~~~~~~
+2.0.0
+~~~~~
 1) [TASK] Drop support for TYPO3 7.6, TYPO3 8.7. Add support for TYPO3 9.5. For TYPO3 7.6 ad 8.7 install version 1.3.
 2) [BUGFIX] Database compatibility.